extract immediate expression parser helpers
